+++ CHANGELOG.md
@@ -143,6 +143,11 @@
 
 * Changed:  $staticFile in StaticResourceUrlHandler is now protected
 
+* Fixed:    Dependency injection static memory issues in HttpClient
+* Changed:  CURLOPT_SSL_VERIFYPEER should be true on all curl requests
+* Changed:  A MultiPartFormDataRequest now provides a MimeDocument as it's payload
+* Fixed:    MimeDocument RFC 1341 compliant boundary header detection
+
 ### 1.1.6
 
 * Fixed:    Header value case bug introduced in 1.1.5